refactor:#172 format util 정리
