+++ src/app/(client)/my-page/components/macronutrient-card.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@/components/ui/typography';
-import { formatNumberWithComma } from '@/lib/utils/format-number-with-comma';
+import { formatNumberWithComma } from '@/lib/utils/format.util';
 import { calculateMacroCaloriesbyGrams, CALORIES_PER_GRAM } from '@/lib/utils/nutrition-calculator.util';
 
 type MacronutrientCardProps = {
