+++ src/app/(home)/utils/calendar.util.ts
@@ -1,5 +1,5 @@
 import { CALENDAR_RANGE_OFFSET, DAY, WEEK } from '@/app/(home)/constants/calendar.constant';
-import { formatDateWithDash } from '@/lib/utils/date.util';
+import { formatDateWithDash } from '@/lib/utils/format.util';
 
 export const isSameDate = (d1: Date, d2: Date): boolean => formatDateWithDash(d1) === formatDateWithDash(d2);
