+++ src/lib/utils/format.util.ts
@@ -1,17 +1,16 @@
-export const formatDateCaption = (date: Date) => {
+/**
+ * 숫자의 천 단위마다 콤마(,)를 붙이는 함수입니다.
+ *
+ * @function formatNumberWithComma
+ * @param {number} number - 변환할 숫자
+ * @returns {string} 천 단위마다 콤마(,)가 포함된 문자열
+ */
+export function formatNumberWithComma(number: number): string {
+  return number.toLocaleString('ko-KR');
+}
+
+export const formatDateToLocaleKR = (date: Date) => {
   return `${date.getFullYear()}년 ${date.getMonth() + 1}월`;
-};
-
-export const formatAmPmKorean = (dateString: string) => {
-  const date = new Date(dateString);
-  return date.getHours() < 12 ? '오전' : '오후';
-};
-
-export const formatTime = (dateString: string) => {
-  const date = new Date(dateString);
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-  return `${hours}:${minutes}`;
 };
 
 export const formatDateWithDash = (date: Date): string => {
@@ -26,7 +25,7 @@ export const formatDateWithDash = (date: Date): string => {
   return formattedDate;
 };
 
-export const convertToTimestamp = (
+export const formatTimestamp = (
   baseDate: Date,
   time: { meridiem: '오전' | '오후'; hours: string; minutes: string }
 ): string => {
@@ -45,4 +44,14 @@ export const convertToTimestamp = (
   const hh = String(hour).padStart(2, '0');
 
   return `${yyyy}-${mm}-${dd} ${hh}:${minutes}:00`;
+};
+
+export const formatTimeWithMeridiem = (date: Date) => {
+  return date.getHours() < 12 ? '오전' : '오후';
+};
+
+export const formatTimeToHHMM = (date: Date) => {
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  return `${hours}:${minutes}`;
 };
